feat: Inserindo pacotes do entity framework e reorganização das pastas de infraestutura
