+++ src/Baltaio.Location.Api/Infrastructure/Persistance/ApplicationDbContext .cs
@@ -1,0 +1,23 @@
+﻿using Baltaio.Location.Api.Domain;
+using Baltaio.Location.Api.Domain.Users;
+using Microsoft.EntityFrameworkCore;
+
+namespace Baltaio.Location.Api.Infrastructure.Persistance;
+
+public class ApplicationDbContext : DbContext
+{
+    public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) 
+        : base(options)
+    {
+    }
+
+    public DbSet<City> Cities { get; set; } = null!;
+    public DbSet<State> States { get; set; } = null!;
+    public DbSet<User> Users { get; set; } = null!;
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.ApplyConfigurationsFromAssembly(typeof(ApplicationDbContext).Assembly);
+    }
+}
+
