+++ src/Baltaio.Location.Api/Program.cs
@@ -13,9 +13,8 @@ using Baltaio.Location.Api.Application.Users.Register;
 using Baltaio.Location.Api.Application.Users.Register.Abstraction;
 using Baltaio.Location.Api.Contracts.Cities;
 using Baltaio.Location.Api.Contracts.Users;
-using Baltaio.Location.Api.Infrastructure;
-using Baltaio.Location.Api.Infrastructure.Users.Authentication;
-using Baltaio.Location.Api.Infrastructure.Users.Persistance;
+using Baltaio.Location.Api.Infrastructure.Authentication;
+using Baltaio.Location.Api.Infrastructure.Persistance.Repositories;
 using Baltaio.Location.Api.OpenApi;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc;
